docs, tests: moving unit tests, adding documentation generation
