+++ .gitignore
@@ -62,3 +62,6 @@ typings/
 
 # vscode IDE
 .vscode/
+
+# documentation folder
+docs/
