+++ jsdoc.conf
@@ -1,0 +1,3 @@
+{
+    "plugins": ["plugins/markdown"]
+}
